move siren dataset to its own file
